✏️  [FEAT] Facade 패턴으로 변경 #14
